sc-7337 GDS User UI: refact certificate registration component
