+++ web/gds-user-ui/src/components/Tables/CertificateRegistrationRow.tsx
@@ -1,20 +1,17 @@
-import React from "react";
 import {
-  Tr,
-  Td,
   Flex,
   Text,
-  Icon,
   IconButton,
   Box,
   Menu,
   MenuButton,
   MenuList,
   MenuItem,
-  useColorModeValue,
-} from "@chakra-ui/react";
+  useColorModeValue
+} from '@chakra-ui/react';
+import { Tr, Td } from '@chakra-ui/table';
 
-import { BsThreeDots } from "react-icons/bs";
+import { BsThreeDots } from 'react-icons/bs';
 
 interface CertificateRegistrationRowProps {
   name: string;
@@ -25,57 +22,56 @@ interface CertificateRegistrationRowProps {
 
 const getBackgroundByStatusCode = (status: string | null) => {
   if (!status) {
-    return "#555151";
+    return '#555151';
   }
 };
 
 const CertificateRegistrationRow = (props: CertificateRegistrationRowProps) => {
   const { section, name, description, status } = props;
-  const textColor = useColorModeValue("#858585", "white");
+  const textColor = useColorModeValue('#858585', 'white');
   return (
     <Tr
       border="1px solid #23A7E0"
       borderRadius={100}
       sx={{
         td: {
-          height: "66px",
-          borderTop: "1px solid #23A7E0",
-          borderBottom: "1px solid #23A7E0",
+          height: '66px',
+          borderTop: '1px solid #23A7E0',
+          borderBottom: '1px solid #23A7E0'
         },
-        "td:first-child": {
-          border: "1px solid #23A7E0",
+        'td:first-child': {
+          border: '1px solid #23A7E0',
           borderLeftRadius: 100,
-          borderRight: "none",
-          textAlign: "center",
+          borderRight: 'none',
+          textAlign: 'center'
         },
-        "td:nth-child(3)": {},
-        "td:last-child": {
-          textAlign: "center",
-          border: "1px solid #23A7E0",
+        'td:nth-child(3)': {},
+        'td:last-child': {
+          textAlign: 'center',
+          border: '1px solid #23A7E0',
           borderRightRadius: 100,
-          borderLeft: "none",
-        },
-      }}
-    >
+          borderLeft: 'none'
+        }
+      }}>
       <Td>
-        <Text fontSize="md" color={textColor} pb=".5rem">
+        <Text fontSize="md" color={textColor} pb=".5rem" data-testid="section">
           {section}
         </Text>
       </Td>
-      <Td minWidth={{ sm: "250px" }} pl="0px">
+      <Td minWidth={{ sm: '250px' }} pl="0px">
         <Flex alignItems="center" py=".8rem" minWidth="100%" flexWrap="nowrap">
           <Text
             fontSize="md"
-            color={"#1F4CED"}
+            color={'#1F4CED'}
             fontWeight="bold"
             minWidth="100%"
-          >
+            data-testid="name">
             {name}
           </Text>
         </Flex>
       </Td>
       <Td>
-        <Text fontSize="md" color={textColor} pb=".5rem">
+        <Text fontSize="md" color={textColor} pb=".5rem" data-testid="description">
           {description}
         </Text>
       </Td>
@@ -83,7 +79,7 @@ const CertificateRegistrationRow = (props: CertificateRegistrationRowProps) => {
         <Text fontSize="md" color={textColor} fontWeight="bold" pb=".5rem">
           <Box
             textAlign="center"
-            fontSize={"sm"}
+            fontSize={'sm'}
             width="96px"
             alignItems="center"
             height="27px"
@@ -91,20 +87,15 @@ const CertificateRegistrationRow = (props: CertificateRegistrationRowProps) => {
             color="white"
             py={1}
             background={getBackgroundByStatusCode(status)}
-          >
-            <Text fontSize={"sm"}> {status || "Incomplete"}</Text>
+            data-testid="status">
+            <Text fontSize={'sm'}> {status || 'Incomplete'}</Text>
           </Box>
         </Text>
       </Td>
 
       <Td>
         <Menu>
-          <MenuButton
-            as={IconButton}
-            icon={<BsThreeDots />}
-            borderRadius={50}
-            bg={"transparent"}
-          />
+          <MenuButton as={IconButton} icon={<BsThreeDots />} borderRadius={50} bg={'transparent'} />
           <MenuList>
             <MenuItem>Edit</MenuItem>
           </MenuList>
